sui-bridge-indexer: remove old bridge indexer
Remove the functionality from `sui-bridge-indexer` for indexing Sui
events since this is obsolesced by the new `sui-bridge-indexer-alt`. The
eth indexer functionality is preserved and the `sui-indexer-builder`,
which is only used by the bridge indexer, is removed and the required
functionality has been inlined into the `sui-bridge-indexer` crate.
