+++ crates/sui-bridge-indexer/Cargo.toml
@@ -28,7 +28,6 @@ sui-bridge.workspace = true
 sui-bridge-schema.workspace = true
 sui-sdk.workspace = true
 sui-json-rpc-types.workspace = true
-sui-data-ingestion-core.workspace = true
 sui-pg-db.workspace = true
 sui-types.workspace = true
 telemetry-subscribers.workspace = true
@@ -36,14 +35,12 @@ tracing.workspace = true
 backoff.workspace = true
 sui-config.workspace = true
 tempfile.workspace = true
-sui-indexer-builder.workspace = true
 
 [dev-dependencies]
 sui-test-transaction-builder.workspace = true
 test-cluster.workspace = true
 hex-literal = "0.3.4"
 diesel_migrations = "2.2.0"
-sui-indexer-builder = { workspace = true, features = ["test-utils"] }
 sui-bridge = { workspace = true, features = ["test-utils"] }
 
 [[bin]]
