+++ crates/sui-bridge-indexer/src/metrics.rs
@@ -2,23 +2,32 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use prometheus::{
-    IntCounter, IntCounterVec, IntGauge, IntGaugeVec, Registry,
-    register_int_counter_vec_with_registry, register_int_counter_with_registry,
-    register_int_gauge_vec_with_registry, register_int_gauge_with_registry,
+    IntCounter, IntCounterVec, IntGaugeVec, Registry, register_int_counter_vec_with_registry,
+    register_int_counter_with_registry, register_int_gauge_vec_with_registry,
 };
-use sui_indexer_builder::metrics::IndexerMetricProvider;
+
+pub trait IndexerMetricProvider: Send + Sync {
+    fn get_tasks_latest_retrieved_checkpoints(&self) -> &IntGaugeVec;
+
+    fn get_tasks_remaining_checkpoints_metric(&self) -> &IntGaugeVec;
+
+    fn get_tasks_processed_checkpoints_metric(&self) -> &IntCounterVec;
+
+    fn get_inflight_live_tasks_metrics(&self) -> &IntGaugeVec;
+
+    fn boxed(self) -> Box<dyn IndexerMetricProvider>
+    where
+        Self: Sized + 'static,
+    {
+        Box::new(self)
+    }
+}
 
 #[derive(Clone, Debug)]
 pub struct BridgeIndexerMetrics {
-    pub(crate) total_sui_bridge_transactions: IntCounter,
-    pub(crate) total_sui_token_deposited: IntCounter,
-    pub(crate) total_sui_token_transfer_approved: IntCounter,
-    pub(crate) total_sui_token_transfer_claimed: IntCounter,
-    pub(crate) total_sui_bridge_txn_other: IntCounter,
     pub(crate) total_eth_bridge_transactions: IntCounter,
     pub(crate) total_eth_token_deposited: IntCounter,
     pub(crate) total_eth_token_transfer_claimed: IntCounter,
-    pub(crate) last_committed_sui_checkpoint: IntGauge,
     pub(crate) backfill_tasks_remaining_checkpoints: IntGaugeVec,
     pub(crate) tasks_processed_checkpoints: IntCounterVec,
     pub(crate) tasks_latest_retrieved_checkpoints: IntGaugeVec,
@@ -28,36 +37,6 @@ pub struct BridgeIndexerMetrics {
 impl BridgeIndexerMetrics {
     pub fn new(registry: &Registry) -> Self {
         Self {
-            total_sui_bridge_transactions: register_int_counter_with_registry!(
-                "bridge_indexer_total_sui_bridge_transactions",
-                "Total number of sui bridge transactions",
-                registry,
-            )
-            .unwrap(),
-            total_sui_token_deposited: register_int_counter_with_registry!(
-                "bridge_indexer_total_sui_token_deposited",
-                "Total number of sui token deposited transactions",
-                registry,
-            )
-            .unwrap(),
-            total_sui_token_transfer_approved: register_int_counter_with_registry!(
-                "bridge_indexer_total_sui_token_transfer_approved",
-                "Total number of sui token approved transactions",
-                registry,
-            )
-            .unwrap(),
-            total_sui_token_transfer_claimed: register_int_counter_with_registry!(
-                "bridge_indexer_total_sui_token_transfer_claimed",
-                "Total number of sui token claimed transactions",
-                registry,
-            )
-            .unwrap(),
-            total_sui_bridge_txn_other: register_int_counter_with_registry!(
-                "bridge_indexer_total_sui_bridge_txn_other",
-                "Total number of other sui bridge transactions",
-                registry,
-            )
-            .unwrap(),
             total_eth_bridge_transactions: register_int_counter_with_registry!(
                 "bridge_indexer_total_eth_bridge_transactions",
                 "Total number of eth bridge transactions",
@@ -73,12 +52,6 @@ impl BridgeIndexerMetrics {
             total_eth_token_transfer_claimed: register_int_counter_with_registry!(
                 "bridge_indexer_total_eth_token_transfer_claimed",
                 "Total number of eth token claimed transactions",
-                registry,
-            )
-            .unwrap(),
-            last_committed_sui_checkpoint: register_int_gauge_with_registry!(
-                "bridge_indexer_last_committed_sui_checkpoint",
-                "The latest sui checkpoint that indexer committed to DB",
                 registry,
             )
             .unwrap(),
