+++ crates/sui-bridge-indexer/src/progress.rs
