+++ crates/sui-bridge-schema/Cargo.toml
@@ -11,7 +11,6 @@ diesel = { workspace = true, features = ["serde_json"] }
 diesel_migrations.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-sui-indexer-builder.workspace = true
 sui-field-count.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
